# fibonacci-calculator/dist/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Fibonacci Calculator",
    "description": "Fibonacci Calculator API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:8000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/fibonacci/{position}": {
      "parameters": [
        {
          "name": "position",
          "in": "path",
          "required": true,
          "description": "Position that we want to request in the Fibonacci sequence",
          "type": "integer"
        }
      ],
      "get": {
        "tags": [
          "Fibonacci"
        ],
        "summary": "Get number at requested position in Fibonacci sequence",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Result": {
      "type": "number",
      "$ref": "#/definitions/Result"
    },
...
  }
}